Issue #15: Improving the whole module experience using streams.
This commit improves the experience with the module by using streams. It
simplifies the input verification by detecting if the stream or options
was used as input. Then, it processed them in 2 different ways: with or
with a given app.

Although the module became longer, I think it is more logical. Also
improving in the error handling and logs.

*	modified:   index.js
- Adding event-related apis, logging and plugin descriptor.
- Adding the array of acceptable formats.
- Execute open will execute opening the file or stream, which is
  detected by the queue file function.
- Using the es.through as used by karma, thus, removing the through2
  dependency.
- Adding the "indexOfIndexOf" method to identify the type of prefix of
  the input.

*	modified:   package.json
- Removing through2 and intalling the other dependencies for the
  enhancements.