Issue #15: Adding last changes after code review
This commit adds the .jshintrc in order to avoid downloading it before
testing. As suggested, it should be in the project.

Refactoring the index to return the errors right away, similar to what
is implemented at
https://github.com/stevelacy/gulp-bump/blob/master/index.js#L20.

Finally, removing the posttest changes from package.json as it is not
required anymore.

*	new file:   .jshintrc
- Downloaded from
  https://raw.githubusercontent.com/stevelacy/gulp-stylus/master/.jshintrc

*	modified:   index.js
- All the refactoring suggested by @stevelacy

*	modified:   package.json
- Removing the posttest.